+++ packages/dash-core-components/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- Fix dynamically disabling and enabling `Interval` components [#436](https://github.com/plotly/dash-core-components/pull/436)
+
 ## [0.42.1] - 2019-01-07
 ### Fixed
 - Fix `dcc.Store` type changes [#427](https://github.com/plotly/dash-core-components/pull/427)